Add files via upload
Suppose that a web server company wants to buy new servers to replace outdated ones and has two options to choose from. There is
a standard model which costs $400, uses 300W of power, takes up 2U of a server rack, and can handle 1000 hits/min. There is also a cutting-edge model, which costs $1600, uses 500W of power, but takes up only 1U, and can handle 2000 hits/min. With a budget of $36,800, 44U of server space and 12,200W of power, how many units of each model should the company purchase in order to maximize the number of hits it can serve every minute?